% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specboots.r
\name{spec.boots}
\alias{spec.boots}
\title{Bootstrapping Spectral Density}
\usage{
spec.boots(
  x,
  nboot,
  spans = NULL,
  kernel = NULL,
  detrend = TRUE,
  demean = FALSE,
  base.stat = median,
  alpha = 0.95
)
}
\arguments{
\item{x}{a time series}

\item{nboot}{the number of bootstrap series to compute}

\item{spans}{specify smoothing}

\item{kernel}{specify kernel}

\item{detrend}{if TRUE, data is detrended first (TRUE as default)}

\item{demean}{if TRUE, data is demeaned first (FALSE as default)}

\item{base.stat}{a function applied to return the baseline statistic of interest (median as default)}

\item{alpha}{a confidence interval (0.95 as default)}
}
\value{
A dataframe of frequencies, lower and upper spectral density bounds.
\itemize{
\item{freq: frequencies at which spectral density are computed}
\item{lwr: the spectral density lower bound at all frequencies}
\item{upr: the spectral density upper bound at all frequencies}
}
}
\description{
This function bootstraps kernel spectral density estimates by resampling the periodogram of the original data object.
The result is a plot of the bootstrap confidence interval and a dataframe that contains the frequencies and
the lower and upper bounds of the estimated spectral density.
}
\examples{
data("AirPassengers")
tsplot(AirPassengers, col=4, lwd=2)
kd = kernel("daniell", c(5,5))
spec.boots(AirPassengers, 1000, kd, periodogram_plot = FALSE)
}
\references{
\itemize{
\item{Franke, J., and W. Hardle. “On Bootstrapping Kernel Spectral Estimates.” The Annals of Statistics, vol. 20, no. 1, 1992, pp. 121–45. JSTOR, http://www.jstor.org/stable/2242153.}
\item{Zoubir, Abdelhak M.. “Bootstrapping spectra: Methods, comparisons and application to knock data”. Signal Process. 90, 5. May, 2010, 1424–1435. https://doi.org/10.1016/j.sigpro.2009.11.030.}
}
}
